z-alpha-ordering

+++ foliage/src/circle/circle.wgsl
@@ -9,14 +9,14 @@ struct Vertex {
     @location(4) layer: f32,
     @location(5) color: vec4<f32>,
     @location(6) ring: f32,
-    @location(7) mip: u32,
+    @location(7) mip: f32,
 };
 struct VertexFragment {
     @builtin(position) position: vec4<f32>,
     @location(0) texture_coordinates: vec2<f32>,
     @location(1) color: vec4<f32>,
     @location(2) ring: f32,
-    @location(3) mip: u32,
+    @location(3) mip: f32,
 };
 @vertex
 fn vertex_entry(vertex: Vertex) -> VertexFragment {
@@ -38,14 +38,14 @@ fn fragment_entry (vertex_fragment: VertexFragment) -> @location(0) vec4<f32> {
         panel_texture,
         panel_sampler,
         vertex_fragment.texture_coordinates,
-        f32(vertex_fragment.mip)
+        vertex_fragment.mip
     ).r;
     if (vertex_fragment.ring != 0.0) {
         coverage = textureSampleLevel(
             panel_ring_texture,
             panel_sampler,
             vertex_fragment.texture_coordinates,
-            f32(vertex_fragment.mip)
+            vertex_fragment.mip
         ).r;
     }
